% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_text.R
\name{search_text}
\alias{search_text}
\title{Search for text in R files}
\usage{
search_text(string, dir = ".", extension = c("R", "Rmd"), pattern = NULL,
  fixed = FALSE, ignore.case = TRUE, deep = FALSE)
}
\arguments{
\item{string}{text string to search for. If fixed = FALSE (the default), regular expressions may be used}

\item{dir}{directory (by default, the working directory)}

\item{extension}{file extension of files to search, by default .R and .Rmd}

\item{pattern}{if you want to search for files with a particular name (takes presedence over extension)}

\item{fixed}{if TRUE, the literal string will be searched for; if FALSE (the default), regular expressions may be used (see ?grep)}

\item{ignore.case}{if TRUE (default), character case will be ignored (i.e., if you search for "Md" the function will search for "md", "mD" and "MD" as well)}

\item{deep}{if TRUE, it will also search in subdirectories}
}
\description{
This function searches for text in text files with (by default) R and Rmd extension.
}
\examples{
search_text("Mod. 1")
search_text("image.plot(", "H:/Documents/seksjon 214/2015_01_FjordBloom/Analyser", fixed = TRUE, deep = TRUE)
# Use fixed = TRUE when the search text contains e.g. parantheses:
search_text("save(", fixed = TRUE)
}
\keyword{string}
